extract JsonUtils in common module
